Selectively call correct backend API in User table
